feat(tailwind): Add custom colors, box shadow, and screen size
This commit introduces custom colors, a box shadow style, and a new screen size to the Tailwind CSS configuration.

- Added primary, secondary, 	ertiary, lack-100, lack-200, and white-100 colors for a consistent color palette.
- Defined a card box shadow style for a subtle, elevated effect.
- Created a new xs screen size at 450px for better responsiveness on smaller devices.

These changes enhance the visual styling and responsiveness of the application.

Additionally, a new utility folder has been created to house utility-related files. The motion.js file within this folder will be used to implement custom motion effects in the future.