feat: Add navbar component
This commit adds a navbar component to the website.

The navbar includes the following features:

A logo that links to the homepage
A menu button that opens a dropdown menu
Links to the different sections of the website
A search bar
A social media icons
The navbar is responsive and will adjust its layout based on the screen size.